Polished AOP question about ProceedingJoinPoint argument
