+++ src/main/java/spring/certification/aop/q009/example/AspectMarkerConfiguration.java
@@ -1,0 +1,11 @@
+package spring.certification.aop.q009.example;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.EnableAspectJAutoProxy;
+
+@Configuration
+@ComponentScan
+@EnableAspectJAutoProxy
+public class AspectMarkerConfiguration {
+}
